+++ catscripts/Technology/Computers/Software/Multimedia_software/WindowsMediaPlayer.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-egrep -i 'Windows(| )Media(| )Player|\bWMP\b' newpages.txt >> WindowsMediaPlayer.txt
+KEYWORDS_WINDOWSMEDIAPLAYER="Windows(| )Media(| )Player|\bWMP\b"
 
-WMP=`stat --print=%s WindowsMediaPlayer.txt`
-
-if [ $WMP -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="WindowsMediaPlayer.txt"
-  export CATNAME="Windows Media Player"
-  $CATEGORIZE
-fi
+  
+  egrep -i "$KEYWORDS_WINDOWSMEDIAPLAYER" newpages.txt >> WindowsMediaPlayer.txt
 
-rm WindowsMediaPlayer.txt
+  WMP=`stat --print=%s WindowsMediaPlayer.txt`
+
+  if [ $WMP -ne 0 ];
+  then
+    export CATFILE="WindowsMediaPlayer.txt"
+    export CATNAME="Windows Media Player"
+    $CATEGORIZE
+  fi
+
+  rm WindowsMediaPlayer.txt
+
+fi
