+++ catscripts/Technology/Computers/Software/Multimedia_software/iTunes.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-egrep -i 'iTunes|\bi Tunes' newpages.txt >> iTunes.txt
+KEYWORDS_ITUNES="iTunes|\bi Tunes"
 
-ITUNES=`stat --print=%s iTunes.txt`
-
-if [ $ITUNES -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="iTunes.txt"
-  export CATNAME="iTunes"
-  $CATEGORIZE
-fi
 
-rm iTunes.txt
+  egrep -i "$KEYWORDS_ITUNES" newpages.txt >> iTunes.txt
+
+  ITUNES=`stat --print=%s iTunes.txt`
+
+  if [ $ITUNES -ne 0 ];
+  then
+    export CATFILE="iTunes.txt"
+    export CATNAME="iTunes"
+    $CATEGORIZE
+  fi
+
+  rm iTunes.txt
+
+fi
