+++ catscripts/Technology/Computers/Software/Operating_systems/Mac_OS_X/Mac_OS_X.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-egrep -i 'Mac(| )OS' newpages.txt >> MacOSX.txt
+KEYWORDS_MACOS="Mac(| )OS|OS(| )X"
 
-MACOS=`stat --print=%s MacOSX.txt`
 
-if [ $MACOS -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="MacOSX.txt"
-  export CATNAME="Mac OS X"
-  $CATEGORIZE
-fi
+  egrep -i "$KEYWORDS_MACOS" newpages.txt >> MacOSX.txt
 
-rm MacOSX.txt
+  MACOS=`stat --print=%s MacOSX.txt`
+
+  if [ $MACOS -ne 0 ];
+  then
+    export CATFILE="MacOSX.txt"
+    export CATNAME="Mac OS X"
+    $CATEGORIZE
+  fi
+
+  rm MacOSX.txt
+
+fi
