+++ catscripts/Technology/Computers/Software/Web_browsers/Webbrowsers.sh
@@ -1,44 +1,56 @@
 #!/bin/bash
 
-egrep -i 'Internet Explorer|\bIE(| )[0-9]{1,2}\b' newpages.txt >> InternetExplorer.txt
-egrep -i 'Firefox' newpages.txt >> MozillaFirefox.txt
-egrep -i 'Google Chrome|Chrome browser' newpages.txt >> GoogleChrome.txt
-egrep -i 'web browser' newpages.txt | egrep -iv 'Chrome|Firefox|Internet Explorer|\bIE(| )[0-9]{1,2}' >> Webbrowsers.txt
+KEYWORDS_FIREFOX="Firefox"
+KEYWORDS_GOOGLECHROME="Google(| )Chrome|Chrome(| )browser"
+KEYWORDS_INTERNETEXPLORER="Internet(| )Explorer|\bIE(| )[0-9]{1,2}\b"
+KEYWORDS_WEBBROWSER="Web(| )browser"
+KEYWORDS_WEBBROWSER_EXCLUDE="$KEYWORDS_FIREFOX|$KEYWORDS_GOOGLECHROME|$KEYWORDS_INTERNETEXPLORER"
+KEYWORDS_WEBBROWSER_OTHER="NCSA(| )Mosaic|Safari(| )browser|Chromium(| )browser|\bDillo|Netscape(| )Navigator|Arachne|\bAmaya\b|NetFront|NetSurf|WebKit|Konqueror|Kazekhase|Mozilla"
+KEYWORDS_WEBBROWSER_ALL="$KEYWORDS_WEBBROWSER|$KEYWORDS_WEBBROWSER_EXCLUDE|$KEYWORDS_WEBBROWSER_OTHER"
 
-IEXPLORER=`stat --print=%s InternetExplorer.txt`
-FIREFOX=`stat --print=%s MozillaFirefox.txt`
-CHROME=`stat --print=%s GoogleChrome.txt`
-BROWSERS=`stat --print=%s Webbrowsers.txt`
-
-if [ $IEXPLORER -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="InternetExplorer.txt"
-  export CATNAME="Internet Explorer"
-  $CATEGORIZE
-fi
+  egrep -i "$KEYWORDS_INTERNETEXPLORER" newpages.txt >> InternetExplorer.txt
+  egrep -i "$KEYWORDS_FIREFOX" newpages.txt >> MozillaFirefox.txt
+  egrep -i "$KEYWORDS_GOOGLECHROME" newpages.txt >> GoogleChrome.txt
+  egrep -i "$KEYWORDS_WEBBROWSER|$KEYWORDS_WEBBROWSER_OTHER" newpages.txt | egrep -iv "$KEYWORDS_WEBBROWSER_EXCLUDE" >> Webbrowsers.txt
 
-if [ $FIREFOX -ne 0 ];
-then
-  export CATFILE="MozillaFirefox.txt"
-  export CATNAME="Mozilla Firefox"
-  $CATEGORIZE
-fi
+  IEXPLORER=`stat --print=%s InternetExplorer.txt`
+  FIREFOX=`stat --print=%s MozillaFirefox.txt`
+  CHROME=`stat --print=%s GoogleChrome.txt`
+  BROWSERS=`stat --print=%s Webbrowsers.txt`
 
-if [ $CHROME -ne 0 ];
-then
-  export CATFILE="GoogleChrome.txt"
-  export CATNAME="Google Chrome"
-  $CATEGORIZE
-fi
+  if [ $IEXPLORER -ne 0 ];
+  then
+    export CATFILE="InternetExplorer.txt"
+    export CATNAME="Internet Explorer"
+    $CATEGORIZE
+  fi
 
-if [ $BROWSERS -ne 0 ];
-then
-  export CATFILE="Webbrowsers.txt"
-  export CATNAME="Web browsers"
-  $CATEGORIZE
-fi
+  if [ $FIREFOX -ne 0 ];
+  then
+    export CATFILE="MozillaFirefox.txt"
+    export CATNAME="Mozilla Firefox"
+    $CATEGORIZE
+  fi
 
-rm InternetExplorer.txt
-rm MozillaFirefox.txt
-rm GoogleChrome.txt
-rm Webbrowsers.txt
+  if [ $CHROME -ne 0 ];
+  then
+    export CATFILE="GoogleChrome.txt"
+    export CATNAME="Google Chrome"
+    $CATEGORIZE
+  fi
+
+  if [ $BROWSERS -ne 0 ];
+  then
+    export CATFILE="Webbrowsers.txt"
+    export CATNAME="Web browsers"
+    $CATEGORIZE
+  fi
+
+  rm InternetExplorer.txt
+  rm MozillaFirefox.txt
+  rm GoogleChrome.txt
+  rm Webbrowsers.txt
+
+fi
